+++ sources/scripts/js/module/Workspace/WorkspaceMemberManagement.js
@@ -1,0 +1,7 @@
+'use strict';
+
+Vaultier.WorkspaceMemberManagementRoute = Vaultier.MemberManagementRoute.extend(
+    Vaultier.WorkspaceKeysMixin,
+    Vaultier.WorkspaceMixin);
+
+Vaultier.WorkspaceMemberManagementController = Vaultier.MemberManagementController.extend({});