+++ sources/static/vaultier/js/loader.js
@@ -59,7 +59,7 @@ ApplicationLoader = {
 
         if (this.loaderEl) {
             if (!duration) {
-                duration = 100
+                duration = 500
             }
             this.loaderElQueue
                 .fadeOut(duration)