+++ sources/static/vaultier/vault.tpl.js
@@ -247,57 +247,53 @@ function program34(depth0,data) {
 Ember.TEMPLATES["Vault/VaultsIndexItem"] = Ember.Handlebars.template(function anonymous(Handlebars,depth0,helpers,partials,data) {
 this.compilerInfo = [4,'>= 1.0.0'];
 helpers = this.merge(helpers, Ember.Handlebars.helpers); data = data || {};
-  var buffer = '', stack1, helperMissing=helpers.helperMissing, escapeExpression=this.escapeExpression, self=this;
+  var buffer = '', stack1, helper, options, helperMissing=helpers.helperMissing, escapeExpression=this.escapeExpression, self=this;
 
 function program1(depth0,data) {
   
   var buffer = '', stack1, helper, options;
-  data.buffer.push("\n    ");
-  stack1 = (helper = helpers['link-to'] || (depth0 && depth0['link-to']),options={hash:{
-    'class': ("vlt-item vlt-vault-item")
-  },hashTypes:{'class': "STRING"},hashContexts:{'class': depth0},inverse:self.noop,fn:self.program(2, program2, data),contexts:[depth0,depth0],types:["STRING","ID"],data:data},helper ? helper.call(depth0, "Vault.index", "", options) : helperMissing.call(depth0, "link-to", "Vault.index", "", options));
+  data.buffer.push("\n        <div class=\"vlt-visual\">\n        </div>\n        <div class=\"vlt-content\">\n            <div class=\"vlt-header\">\n                <h4>");
+  data.buffer.push(escapeExpression((helper = helpers['dot-dot-dot'] || (depth0 && depth0['dot-dot-dot']),options={hash:{
+    'value': ("name"),
+    'height': (60)
+  },hashTypes:{'value': "ID",'height': "INTEGER"},hashContexts:{'value': depth0,'height': depth0},contexts:[],types:[],data:data},helper ? helper.call(depth0, options) : helperMissing.call(depth0, "dot-dot-dot", options))));
+  data.buffer.push("</h4>\n\n                <!--\n                <div class=\"vlt-icon\">\n                    <img src=\"/static/icons/vault/ACROBITS.png\">\n                </div>\n                -->\n            </div>\n            <div class=\"vlt-description\">\n                ");
+  stack1 = helpers['if'].call(depth0, "description", {hash:{},hashTypes:{},hashContexts:{},inverse:self.program(4, program4, data),fn:self.program(2, program2, data),contexts:[depth0],types:["ID"],data:data});
   if(stack1 || stack1 === 0) { data.buffer.push(stack1); }
-  data.buffer.push("\n");
-  return buffer;
-  }
-function program2(depth0,data) {
-  
-  var buffer = '', stack1, helper, options;
-  data.buffer.push("\n        <div class=\"vlt-header\">\n            <div class=\"vlt-icon\">\n                <img src=\"/static/vaultier/images/icon-vault-dark-blue.png\">\n            </div>\n            <div class=\"vlt-title\">\n                <h3>\n                    ");
-  data.buffer.push(escapeExpression((helper = helpers.ellipsis || (depth0 && depth0.ellipsis),options={hash:{},hashTypes:{},hashContexts:{},contexts:[depth0,depth0],types:["ID","INTEGER"],data:data},helper ? helper.call(depth0, "name", 35, options) : helperMissing.call(depth0, "ellipsis", "name", 35, options))));
-  data.buffer.push("\n                </h3>\n\n            </div>\n        </div>\n        <div class=\"vlt-body\">\n            ");
-  stack1 = helpers['if'].call(depth0, "description", {hash:{},hashTypes:{},hashContexts:{},inverse:self.program(5, program5, data),fn:self.program(3, program3, data),contexts:[depth0],types:["ID"],data:data});
-  if(stack1 || stack1 === 0) { data.buffer.push(stack1); }
-  data.buffer.push("\n        </div>\n        <div class=\"vlt-footer\">\n            <div class=\"vlt-footer-item help-block pull-left\">\n                ");
+  data.buffer.push("\n            </div>\n            <div class=\"vlt-footer\">\n                ");
   data.buffer.push(escapeExpression((helper = helpers.printUser || (depth0 && depth0.printUser),options={hash:{
     'ellipsis': (12),
     'prefix': ("Created by:")
   },hashTypes:{'ellipsis': "INTEGER",'prefix': "STRING"},hashContexts:{'ellipsis': depth0,'prefix': depth0},contexts:[depth0],types:["ID"],data:data},helper ? helper.call(depth0, "created_by", options) : helperMissing.call(depth0, "printUser", "created_by", options))));
-  data.buffer.push("\n            </div>\n            <div class=\"vlt-footer-item help-block pull-right\">\n                ");
+  data.buffer.push("\n                ");
   data.buffer.push(escapeExpression((helper = helpers.printAgo || (depth0 && depth0.printAgo),options={hash:{
     'prefix': ("Latest modification at:")
   },hashTypes:{'prefix': "STRING"},hashContexts:{'prefix': depth0},contexts:[depth0],types:["ID"],data:data},helper ? helper.call(depth0, "updated_at", options) : helperMissing.call(depth0, "printAgo", "updated_at", options))));
-  data.buffer.push("\n            </div>\n            <div class=\"clearfix\"></div>\n        </div>\n        <div class=\"clearfix\"></div>\n    ");
+  data.buffer.push("\n            </div>\n        </div>\n    ");
   return buffer;
   }
-function program3(depth0,data) {
+function program2(depth0,data) {
   
   var buffer = '', helper, options;
+  data.buffer.push("\n                    ");
+  data.buffer.push(escapeExpression((helper = helpers['dot-dot-dot'] || (depth0 && depth0['dot-dot-dot']),options={hash:{
+    'value': ("description"),
+    'height': (80)
+  },hashTypes:{'value': "ID",'height': "INTEGER"},hashContexts:{'value': depth0,'height': depth0},contexts:[],types:[],data:data},helper ? helper.call(depth0, options) : helperMissing.call(depth0, "dot-dot-dot", options))));
   data.buffer.push("\n                ");
-  data.buffer.push(escapeExpression((helper = helpers.ellipsis || (depth0 && depth0.ellipsis),options={hash:{},hashTypes:{},hashContexts:{},contexts:[depth0,depth0],types:["ID","INTEGER"],data:data},helper ? helper.call(depth0, "description", 250, options) : helperMissing.call(depth0, "ellipsis", "description", 250, options))));
-  data.buffer.push("\n            ");
   return buffer;
   }
 
-function program5(depth0,data) {
+function program4(depth0,data) {
   
   
-  data.buffer.push("\n                No description given\n            ");
+  data.buffer.push("\n                    No description given\n                ");
   }
 
-  stack1 = helpers['if'].call(depth0, "id", {hash:{},hashTypes:{},hashContexts:{},inverse:self.noop,fn:self.program(1, program1, data),contexts:[depth0],types:["ID"],data:data});
+  data.buffer.push("<div class=\"vlt-vault-item pull-left\">\n    ");
+  stack1 = (helper = helpers['link-to'] || (depth0 && depth0['link-to']),options={hash:{},hashTypes:{},hashContexts:{},inverse:self.noop,fn:self.program(1, program1, data),contexts:[depth0,depth0],types:["STRING","ID"],data:data},helper ? helper.call(depth0, "Vault.index", "", options) : helperMissing.call(depth0, "link-to", "Vault.index", "", options));
   if(stack1 || stack1 === 0) { data.buffer.push(stack1); }
-  data.buffer.push("\n\n");
+  data.buffer.push("\n</div>\n\n");
   return buffer;
   
 });