+++ changelog.rst
@@ -9,6 +9,8 @@ Release 0.7.0
 * [feature] #195 Search usability - make search window more responsive
 * [bugfix] Display message "No matches found" when search box contains no results
 * [bugfix] Search box remembered old names.
+* [feature] #199 Redesign cards and vault item
+* [feature] #180 clever footer
 
 Release 0.6.5
 -------------